backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# LLM and AI
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.1
langgraph==0.0.20
openai==1.3.7
anthropic==0.7.8

# Vector database and embeddings
lancedb==0.4.3
sentence-transformers==2.2.2

# Data processing and analysis
pandas==2.1.3
numpy==1.25.2
plotly==5.17.0

# Audio processing
whisper==1.1.10
pydub==0.25.1

# Security and authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2
schedule==1.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0 